Step 10 (self check): 1. move game-files to the /game directory (new package creating), made all *Game-constructors - public, otherwise errors occurred during compilation 2. rename Cli-class methods 3. code formating
