--- app/src/main/java/hexlet/code/Game.java
@@ -1,8 +1,0 @@
-package hexlet.code;
-
-public interface Game {
-
-    void outDescription();
-    void play();
-    int getHotPoint();
-}
