+++ app/src/main/java/hexlet/code/game/Game.java
@@ -1,0 +1,8 @@
+package hexlet.code.game;
+
+public interface Game {
+
+    void outDescription();
+    void play();
+    int getHotPoint();
+}
